Avoid invoking `Type.create` in `Environment.Builder`
Summary:
The Problem
==========

`Type.Cache` is broken.

This cache is used by `Type.create`, whose result depends on the alias map that gets passed in. In other words, we should really cache the result of `Type.create` separately for different alias maps. Unfortunately, this is not what we do today. `Type.create` gets invoked with all kinds of different alias map, and we make no effort to distinguish them:
- At alias registration time, `Type.create` gets called with an alias map that keeps mutating. We get around this issue with the `enable`/`disable` trick, but this feels more of a hack than a proper solution.
- At type checking time, although the alias map in the environment is no longer mutated, we still did not call `Type.create` with it consistently: There are places where we simply invoke `Type.create` with an alias map deliberately set to empty.

Obviously this is bad because it makes `Type.create` would return incorrect result when the same expression gets parsed, but under different alias environment.

The Solution
==========

Separating the caches out for different alias map is not possible since the alias map is passed in as a closure.

The short term solution is to manually remove all invocations to `Type.create` with empty alias map when `Type.Cache` is enabled. This diff tries to do that for one callsite. Subsequent diffs are planned for other callsites. If it is not possible to remove the usage, we should at least make them skipping the cache.

The long term solution is to change the API and try to make it impossible to invoke `Type.create` with inconsistent alias maps. Due to its dependency on the alias map, it might make more sense to move this API out of `Type` module and tie it to the environment.

Reviewed By: dkgi

Differential Revision: D15110744

fbshipit-source-id: 0a9f169b87dfb4b4ef323267389ebe5cc327a93a