Support new AST in Statement.Define
Summary:
Converting all of Statement.Define's helper functions to handle both new and old AST; whatever the input is fine. Working towards being able to ungate typecheck and have all these ancillary functions "just work".

`implicit_attributes` does some AST generation that is not gated on a match that verifies what the existing AST is... this is tricky because we can't infer which kind of AST to generate. Sadly, this means `implicit_attributes` needs to accept a convert flag, making things more complicated than expected. Thought about tracking whether an AST object is converted already with some field in the AST, but I think that would be really invasive and hopefully sticking with flags will prove to be the easier interim situation.

Reviewed By: sinancepel

Differential Revision: D15035145

fbshipit-source-id: ac4b57e1873ea4c6856ab522a41e501809a06a8c