Creation of reference from name should return optional
Summary:
We used to fail if the name passed in contained some non-identifier base expression, but really we should support the option to return an optional rather than throwing.

Some cleanup would be nice for the old use cases that checked if a name was simple before turning it into a reference, but just make both available for now to unblock access fold rewrite.

Reviewed By: sinancepel

Differential Revision: D15069907

fbshipit-source-id: 4aefbd98a39358aca15b3811f68aea0d59bfae78