Split `Analysis.Environment.register_aliases` into collection and resolution
Summary:
The plan is to move the fixpoint computation of alias resolution (i.e. the `fixpoint` function inside `resolve_aliases`) to `Service.Environment` so that we can parallelize it.

As a first step, separate out collection and resolution so that we can focus on resolution only in subsequent diffs.

Added a utility module `AliasEntry` to replace the tuple we used before.

Reviewed By: dkgi

Differential Revision: D15220578

fbshipit-source-id: e77e3470668acc5857aebf0b4dc7fb9511281cd0