Avoid need to reset namespace for typecheck
Summary:
Namespaces are integers attached to type variables in order to distinguish between different variables with the same name.  The common case this seeks to handle is things like:
```
x: typing.Tuple[typing.List[int], typing.List[str]] = [], []
```
Because they come from the same constructor, the escaped type variables from each empty list have the same name, but they need not be the same value.  We achieve this with a process-global fresh namespace counter.

The initial implementation had to reset the fresh namespace number on every statement in order to avoid infinite loops due this dependency on global state.

Because we prevent escaped type variables from making it into the resolution, the only problematic area was errors.  We can avoid resetting by instead making the Set of errors use a comparator that is insensitive to differences between namespaces.

We don't want to make this a global behavior because the distinction between namespaced variables is the whole reason they were introduced.

Footnote:
This could potentially be more performant through making a version of compare which only patched the compare when it actually hit a variable/variable comparison rather than transforming them pre-emptively.  This doesn't seem necessary right now, but this architecture could accommodate it if it ends up being important.

Reviewed By: grievejia

Differential Revision: D15100261

fbshipit-source-id: 7459a8d26c8cf8ed6f077376546fa54b7cd69bbe