Avoid invoking `Type.create` in `expression_contains_any`
Summary:
See D15110744 for rationales on why we want to remove `Type.create` invocation here.

The original functionality (which was introduced in D13983363) can be entirely replaced by a syntactical check on the given expression. Note that we need to make sure to also count things like `list`, `dict`, etc. as containing literal any.

Reviewed By: dkgi

Differential Revision: D15230490

fbshipit-source-id: 5e811c450cd9e29864129e24e90ab8c071849b4f