Purify `Analysis.Environment.resolve_aliases`
Summary:
This was originally done for alias resolution parallelization: we detach single-round `resolve_aliases` with `fixpoint` so that it is possible to lift `fixpoint` to somewhere else than `analysis/environment.ml`.

Now that we've abandoned the idea of parallelization, this is more of a cosmetic change that prepares for my next idea of speeding things up.

Reviewed By: dkgi

Differential Revision: D15249362

fbshipit-source-id: 18d7a988028b96e372b7c5f919e6100711e240ee