Do not gate alias resolution with TypeOrder existence check
Summary:
There were some discussions about this in D15052310. Unfortunately I was not knowledgable enough back then to understand what Sinan Cepel and Mark Mendoza were talking about :P

My observation here is that the `TypeOrder.contains` check should not exist. It not only hurts perf but also semantically incorrect as well. Consider this example:
```
class A:
  x: int

class B:
  pass

A = B
def foo(a: A) -> int:
  return a.x
foo(A())
```
If we run it, we get `AttributeError: 'B' object has no attribute 'x'`. But if we run `pyre` on it, we get the confusing `Incompatible parameter type: Expected "A" for 1st anonymous parameter to call "foo" but got "B"`. The reason is that in alias registration, we artificially exclude registering any alias that shares the same name with a class. Therefore the type alias definition `A = B` never gets processed and `foo` would be inferred to have type `A -> int`, which is not what we want.

This diff removes the artificial exclusion we mentioned before, which would give us the undefined attribute x error back. However, this is not perfect either. Consider a variant of the previous example:
```
class A:
  x: int

class B:
  pass

def foo(a: A) -> int:
  return a.x
A = B
foo(A())
```
Ideally we'd want a parameter mismatching error to be reported on the function invocation. But it's almost impossible within our current framework as we treat alias resolution as a flow-insensitive procedure while our type checking is flow-sensitive. At the moment we have to live with what we get :(

Reviewed By: dkgi

Differential Revision: D15263925

fbshipit-source-id: 5adcd058d612f5530e19cb091d7c01fcbe37d314