hide TypeConstraints solution type
Summary:
Clients of the TypeConstraints module need not be aware of the underlying type of the solution object returned by the solve functions.
For example, it would be reasonable to convert the representation from a Type -> Type map to a Type.Variable -> Type map, and that refactor would be much simpler if we already abstract the interface.
More importantly in order to support variadic type variables we need to give solutions that aren't of the form Type -> Type, and we'd like to hide that.

Reviewed By: dkgi

Differential Revision: D15287207

fbshipit-source-id: 2c8ddf401d324514a315ed8530707f3d531b3137