start catching errors for stub defines
Summary:
We should still be validating the types being declared in functions with ellipses as the body.

The canonical case where we want this is as follows: "def foo(x: Otpinal[int]) -> int: ... # meant to write Optional".

As a side effect, our stub checking should give MUCH more signal now.

Reviewed By: grievejia

Differential Revision: D15241247

fbshipit-source-id: d401f0f8446fa5149c73ea0fbcae9283e35468fc