don't try to analyze the .resource_cache
Summary:
When pyre is called on a directory, we shouldn't attempt to analyze any handles don't correspond to a path (local_root/handle), as that's the definition of what Pyre should analyze.

The practical case here is that typeshed can live under the project root, and we're hoping to analyze stubs. Let's not throw errors on typeshed stubs for users who don't control them.

Reviewed By: dkgi

Differential Revision: D15240655

fbshipit-source-id: edc994cc69fcdcaae1f7d5b4e27c2975867b465e