Reintroduce anonymous parameters
Summary:
D12945788 removed anonymous parameters in order to avoid an anti-pattern where we were folding over parameters in an unsound manner.
However, in order to reasonably support protocols, we need support for __parameters being treated as anonymous.  Doing so would be possible with the current model but would be again leaning on string based serialization of data which ought to be a first class data structure.  This is problematic for perf and is bug-prone.  Especially with the upcoming addition of runtime support for "positional-only" parameters in the runtime, it seems like this is the time for first-class support.

Working through this I found a variety of issues with override compatibility checking, which I cataloged in T44178876.

Reviewed By: dkgi

Differential Revision: D15248893

fbshipit-source-id: b664d8d8d613252c8333c7c7caebe2de3dd6c008