Wrap config paths in os.path.expanduser (#166)
Summary:
Intended to fix #165

 ---

The idea was to cover every spot a path can be supplied via config file, though some of these places may not make much sense. For instance, I'm not familiar enough with Pyre to know if it does in both root and subdirectory of a `SearchPathElement` dictionary.

I haven't used pyre a ton yet, but I pretty quickly found a need for this in at least `typeshed`, `source_directories`, and `binary`. Open to any sorts of feedback on this!

 ---

This is opened as a draft because I would like to be open to feedback at any stage since I'm so new to Pyre, so as such there's still a little bit to do before this is ready for any final reviews/merging.

Still to-do:

- [x] Discuss whether it makes sense on each of these fields or anywhere else in Pyre
- [x] Patch `os.path.expanduser` where necessary to pass preexisting tests
- [x] Add new tests to each relevant field
- [ ] Update documentation
- [ ] Lint
Pull Request resolved: https://github.com/facebook/pyre-check/pull/166

Reviewed By: grievejia

Differential Revision: D15248528

Pulled By: sinancepel

fbshipit-source-id: 6ecbd6140b5d51d60d72ee532aec6b2ef0138977