Do not force-set MACOS_DEPLOYMENT_TARGET on dev mode
Summary:
The `MACOS_DEPLOYMENT_TARGET` env var was set in D7986807 for old MacOS version support. The problem it solved was about pypi deployment, but it also force-set the version in both `setup.sh` and `Makefile` on every Mac, even on dev machines, which lead to tons of "Library XXX was built for newer version of os but linked with an older version" linker warnings.

We remove the env var definition in `setup.sh` and `Makefile` here and only keep it in our deployment build script `build-pypi.sh`. I also removed `OCAMLFIND_IGNORE_DUPS_IN` in our `Makefile` as we no longer rely on `ocamlfind` to build `hack_parallel`.

Reviewed By: sinancepel

Differential Revision: D15302873

fbshipit-source-id: 86b2f87f5daa0e481a59bc56ee14b47b7ee78ec5