support decoding pairs in server decode queries
Summary:
This change adds support for decoding a pair of values given a key, and computing whether the OCaml representation of this pair corresponds to the same value for the definition of equality that we care about, rather than just comparing the bits under the hood when reporting discrepancies.

To test, this diff changes the structure of assert_decode to test both the initial decoding case, as well as the `decode_pair(key, value, value)` case.

Reviewed By: dkgi

Differential Revision: D15311778

fbshipit-source-id: dd97af6a7ef260e947491d6e928f1a5d128f975b