Stop reverting export logic if resulting module is non-existent
Summary:
This check & revert was initially introduced in D13744353, but it seemed like a hacky solution that wasn't directly targeting the issue (qualification). The test that was added in this diff still passes today with this check removed.

Turns out this is also breaking our handling of placeholder stubs that are not leaf nodes; i.e., accessing some submodule of a module that is a placeholder triggers this check that the resulting resolved export does not exist and reverts back to the un-exported reference.

Trying to check that un-exported reference causes unexpected UndefinedAttribute errors.

Reviewed By: grievejia

Differential Revision: D15302838

fbshipit-source-id: d4609bd1d2254292cb1fe454c2e535e2f6f9fe3a