Terminate with errors on callee
Summary:
This is a fix targeted at errors like:
```
+distillery/accounts/create_account_service.py:179:22 Call error [29]: typing.Union[AppPhoneRegService, typing.Undeclared] is not a function.
```
It's not meaningful to error on the rest of the call if the callee base has an undefined name or undefined attribute error on it already. This was implemented before in the access fold and has been translated to forward_expression for names, but not calls.

Looks like the underlying behavior of the errors is actually pre-existing, for example, on the current distillery master version you can run:

```
def foo() -> None:
    if herp:
        x = 1
        if derp:
            pass
    reveal_type(x)
```

And you will get not one but TWO reveal type errors:
```
test_prop.py:48:4 Revealed type [-1]: Revealed type for `x` is `unknown`.
test_prop.py:48:4 Revealed type [-1]: Revealed type for `x` is `Union[typing_extensions.Literal[1], typing.Undeclared]`.
test_prop.py:48:16 Undefined name [18]: Global name `x` is not defined, or there is at least one control flow path that doesn't define `x`.
```

Tried to investigate a bit... here is a CFG of this pattern (but that looks normal): https://our.intern.facebook.com/intern/graphviz/?paste=64188838

Guessing this has something to do with the CFG fixpoint, since `reveal_type(x)` winds up passing through forward statement twice and the error thrown each time is different.

Reviewed By: grievejia

Differential Revision: D15295874

fbshipit-source-id: 873d155455d3ce2dddea4cc51067b58cb201cdb4