Refine terminating error logic
Summary:
Undefined attribute or name errors might come from places like the parameters of a call in the base or callee; we don't actually want to terminate further erroring on the access chain if they exist but we still have a known resolved type for the base or callee.

Gets rid of some bad extraneous ignore errors on master.

Reviewed By: dkgi

Differential Revision: D15322979

fbshipit-source-id: 30959cf4897541a31b47f16dd241538921377f0f