Recognize setUp as valid test constructor
Summary:
Our original logic seemed correct but we weren't taking into account any of the non `__init__` constructors in the case of tests. Was throwing false positives, as in: D15338562

Move this logic to annotatedClass/Define so that we can look up superclasses, determine if we're in a test, then determine whether something is a constructor.

Reviewed By: dkgi

Differential Revision: D15345250

fbshipit-source-id: ff3ce67968d0524e70259f76fb02af768ea2153a