keep errors in a map of (location, kind) -> error instead of a set
Summary:
We were running into inconsistent errors getting reported by Pyre due to blindly unioning all errors we had and joining them per define after-the-fact. This was problematic when we had different fixpoint iterations generate incompatible errors, which can happen in loops depending on which branch is taken first.

Instead of this ad-hoc joining, enforce the invariant that we can only have one error per (location, error code) pair. The implication is that we now have to join errors thrown for the same location.

For a simple example of the limitations of the previous approach, consider the following (the individual case could be fixed, but the general problem of joining errors after the fact remains):

```
x: str = ""
for y in [1,2,3]:
  reveal_type(x)
  x: int = 2
```

Before:
```
$ pyre check
 ƛ Found 2 type errors!
a.py:3:4 Revealed type [-1]: Revealed type for `x` is `str`.
a.py:3:4 Revealed type [-1]: Revealed type for `x` is `int`.
```

After:
```
$ pyre check
 ƛ Found 1 type error!
a.py:3:4 Revealed type [-1]: Revealed type for `x` is `typing.Union[int, str]`.
```

Reviewed By: shannonzhu

Differential Revision: D15370236

fbshipit-source-id: ecad3a1a2c5a56a1d686d3724e52bacf00be1cae