Resolve typeshed search path at client side
Summary:
Goal
====

We want to systematically tackle the dependency issue for pyre incremental.

The fundamental question we would like to have a good answer here is that whenever something changes in a project, what needs to be recomputed. We can break down this question into two sub-parts: (1) How to correctly figure out what has changed, and (2) Given the changeset, how to do the recomputation.

This diff (and probably the followups) tries to answer question (1). Answering question (2) is the next item on my todo list.

The Problem
==========

Let's first look at how we determine which module(s) has changed for a given project. This doesn't sound like a hard problem at first glance: `watchman` already gives us the list of changed files on the filesystem. Just translate these file paths to module names seems to do the trick?

However, what makes this problem tricky is the fact that one module name may correspond to multiple potential files. Various reasons may contribute to this. One example is the simultaneous existence of `src/a.py` and `src/a/__init__.py`. Another example is the simultaneous existence of `ext/a.py` and `src/a.py`, where `ext/` is in an external search path and `src` is in the project source directory.

It won't be too hard to see that if we do not have a deterministic and consistent rule to pick the right file for a given module, then it's going to be hopeless to correctly compute the file changeset. Unfortunately, the rule we have today is that we sort the source files according to certain order and pick the first one of the that list, and the order we choose to sort the files depends on filesystem iteration, which [is not guaranteed to be in any particular order](http://man7.org/linux/man-pages/man3/readdir.3.html).

The Solution
==========

We want to eliminate our dependency on filesystem behavior. This begins with the typeshed: previously we do not put typeshed subdirectories in any fixed order. This diff tries to enforce one: `stdlib` comes before `third_party`, and later python version comes before earlier python version. Followup diffs will try to fix file listing orders inside each search path in the backend.

Reviewed By: sinancepel

Differential Revision: D15401856

fbshipit-source-id: c39ddbeb2abd103e6e0158adc60884b801122e9d