functorize type checking state
Summary: This will allow us to store global information outside of the state. I've illustrated this here by moving the `Define.t` out of the state. This does not change during the analysis so there's no reason to propagate this through the fixpoint. With this setup we can also move the error map etc. out as well and minimize the state to things that are affected by the propagation.

Reviewed By: grievejia

Differential Revision: D15405412

fbshipit-source-id: a2b05907199df4d00462e4978a92c36604324af7