Fix imports of modules
Summary:
We were false-positive erroring on certain kinds of chain imports, when a module is the object of indirect import.

For example:

a.py
```
class Foo: ...
```

b.py
```
from . import a

# a.Foo is defined here
```

c.py
```
from .b import a

# a.Foo is NOT defined here - it resolves to Unknown and throws InvalidType errors when used as an annotation.
```

When looking at type resolution, we see this is because we are resolving `a.Foo` into `b.a.Foo`, which does not exist. Turns out this is because the `b.a -> a` alias is generated, but then not added to the type order.

Turns out this is because we try to calculate dependencies when registering aliases, and won't register an alias if we deem its value (where from a import b means a is the target and b is the value) to be unresolved. We used to only check whether the value is a type stub or is in the type order, but we need to also check if the value is a module. If so, we can register the alias.

Reviewed By: grievejia

Differential Revision: D15424036

fbshipit-source-id: b847a8ba1d8b96f4c90ec9eeea966f1ecf1ae003