Always double check against protocols in mismatch_with_any
Summary:
We defer handling of Any until the error filtering phase to keep our type lattice free of loops.
As a performance optimization, we only re-run checks on errors which mismatch including an Any in the actual type.
However, when mismatching against protocols, the dependency on Any can be hidden inside of the class that we were comparing, and so can slip through the check.

Reviewed By: grievejia

Differential Revision: D15437477

fbshipit-source-id: e26ce7e2d1e985e854832c574660ecc3ba17ca5e