Make instantiate_predecessors_parameters independent of file load order
Summary:
During our root cause analysis of a bug that arose from inconsistent ordering of the edges from bottom, we realized that all of the backedges were inconsistently ordered, which meant that meet, which does an order sensitive traversal of the backedges, could be inconsistent.

The situation that triggers this inconsistency is pretty convoluted (see the giant test), but it's worth fixing.

Reviewed By: dkgi

Differential Revision: D15461537

fbshipit-source-id: f4ce7353e1bf5010461393a367d71b41da6d7abe