Turn on warnings on detached or ambiguous docstring
Summary:
To prepare for `ocamlformat` adoption, we need to start tracking and getting rid of warning 50 in our codebase, as that's one of the [preconditions `ocamlformat` requires](https://github.com/ocaml-ppx/ocamlformat#overview).

We could, in principle, disable the check in `ocamlformat` with a command-line flag (config file doesn't work as they claimed). But since the fix is not too bad I'd rather do the "right thing" here.

Reviewed By: mrkmndz

Differential Revision: D15471732

fbshipit-source-id: d118fc85ed6ff483047e9e42d1f289760adcd17d