Add structural support for special methods
Summary:
makes me a bit sad to have to add another field here, but I think this is the best way to achieve the end of differentiating between attributes that were created from us preprocessing syntactic sugar vs. attributes that were explicitly spelled out with dot access notation.

for more context on special methods and what is trying to fix: T35601774, D14413641

note: this diff is a no-op; the field 'special' is always set to false and never used. stacking a diff that introduces the logic on top.

Reviewed By: dkgi

Differential Revision: D15461087

fbshipit-source-id: ddc8a96d3c748e81d4f500212f3ce1a6f01535ae