Build out general interface for constraint solving
Summary:
The other half of preparing for variadic type variables comes from generalizing the other source of complexity re: type variables --- constraint solving.

The insight this diff relies on is that just as every type variable has a corresponding domain, we can define intervals on those domains.  An interval refers to the subset of the variables domain that we have determined its solution must belong to.

The core of our constraint solving approach comes from intersecting intervals as we build up constraints, and then selecting members from those intervals when we are solving.

The interval for unaries is just a pair of types, and intersecting those is a simple meet and join.  However, we can define these operations for other domains, rather than the type domain.

With these interfaces for an interval corresponding to a kind of variable, this diff defines a functor to build an interval container, which is what the whole constraints object is built out of.

We then needed to redesign the core functions to delegate the right calls to these underlying containers.

Reviewed By: grievejia

Differential Revision: D15447692

fbshipit-source-id: 562e429ce55af2f3eb9d0988bf91c0eded61c3ca