Build out a general interface for type variable "kinds"
Summary:
In preparation for variadic type variables, this diff builds out a general interface for what makes a kind of type variable, at least with regards to the "global" (with regards to an entire type) transforms necessary to enforce the right behavior around namespacing, marking as bound/free/escaped etc.

The core interface here bears some explaining:
For every kind of variable, we have an ocaml type that corresponds to the variable itself, and an ocaml type that corresponds to the domain of the type variable, or the set of values which could be substituted for the variable.
For our familiar unaries, those correspond to the type inside `Type.Variable` and `Type.t` respectively.
For callable parameter variadics it will correspond to their new record type and `Type.Callable.parameters`.
For list variadics it would be another new record and some kind of variant that either contained a list of types or another list variadic.

Beyond those types, we need to define operations on those types that check its state, mark it as bound or escaped, mark it with a namespace, or dequalify its identifier.
We also need to define an "Any" value within its domain, which corresponds to the behavior we have come to expect from typing.Any.
The last thing in the interface is a self function which yields the value in the domain corresponding to the variable.  I would appreciate suggestions for a potential better name for this function.

With this interface, this diff defines a functor which builds the familiar "global" transformations we have come to rely on, for a given type variable kind.  This means that composing n of these transforms take n traversals over the type.  There are ways to compose these together in one pass, but i think the code ends up significantly more confusing and functor-y, and I'm not convinced it will have a worthwhile perf impact.  Regardless, with this interface we will be able to move to that model if we want to.

Reviewed By: shannonzhu

Differential Revision: D15447677

fbshipit-source-id: 03f7ccf8c54d3f543c7876a33ca57c5f412dd096