Variadic type variables for callables
Summary:
The first application of the newly generalized interfaces for our variable transforms is callable parameter variadics.
# What these are
Type variables which bind to the parameters field of a callable type.  In the lingo of the preceding diff, that is its domain.

# Why are these necessary
Lots of python decorators have the behavior of altering the return type of the callable without changing the parameter type
It may seem possible to type these with variadics that capture *args and **kwargs separately.
However this runs into the problem of the fact that we're trying to squeeze three categories of parameter, positional-only, keyword only and positional-or-keyword into these two groups, which necessarily loses resolution.
Therefore we need to capture all of these into a single variable, the callable parameter variadic.

# Current Limitations
This diff doesn't surface the ability to create these variables to the user yet.  This was the only separable functionality I found, so this is just split into a follow up diff

The current implementation of the interval intersect for these parameter objects only accepts exact matches.
This means that binding `{Named(x, int)}` and `{Named(x, float)}` to the same callable parameter variadic fails, though we could soundly solve that variable to `{Named(x, int)}`.  This is because most decorators only take one function, so this feature can be excluded from the initial MVP.

A deeper limitation is that functions with a variable as their parameters cannot be called.  This means that we can't really typecheck the bodies of the targeted decorators.  I think this is a reasonable limitation for two reasons
* Many more people depend on decorated functions than write decorators.  Prioritizing their experience seems reasonable
* We could build a system to allow calling them in another iteration.  We could imagine a pair of operators which represent all possible pairs of *args and **kwargs that could come from a given callable parameter variadic, and then require that you call the function with and only with those special *args and **kwargs.

Reviewed By: dkgi

Differential Revision: D15339143

fbshipit-source-id: 6cf21457bbf3b2d82cd576077006925f15744f59