Allow for declaration of parameter variadics
Summary:
In order to make callable parameter variadics available to a user, we need to provide a mechanism for declaring them.
This is done for normal type variables through alias resolution, but that mechanism won't quite work for us here because these variables aren't types.
However, the machinery and storage of aliases still seems appropriate, so we can make it work by widening what the alias table contains: `TrueAlias`s for what were once there, and a separate variant for these variable declarations.

We can store it early, bypassing the fixpoint because these variables can't contain any other types, since they have no constraints.

In truth, this should also be how normal type variables are handled, as the declaration syntax shouldn't be allowed inline, as we currently accept
```
def identity(x: typing.TypeVar("T")) -> typing.TypeVar("T"):
  return x
```
which ought to be rejected.

This is not quite usable in production yet, as we need to get it added to the `typing_extensions` package to work at runtime.
However, this workaround does in fact work:
```
from typing import TYPE_CHECKING
if TYPE_CHECKING:
    from typing_extensions import CallableParameterTypeVariable
else:
    def CallableParameterTypeVariable(name):
        return ...
```

Reviewed By: shannonzhu

Differential Revision: D15447745

fbshipit-source-id: 0f31d8f31ae3450a523dac2c95be6226a2f4037f