Expand traces by leaf kind; Keep leaf callable in issue instance
Summary:
IssueInstances are emitted based on matching source kind to sink kind. Determining the next hop when traversing traces only require that the TraceFrame's leaf kind matches the leaf kind of the IssueInstance. The actual leaf name is not needed.

Storing the leaf kind in TraceFrameLeafAssoc (instead of every reachable leaf name) is sufficient and removes redundant assocs.

The leaf name is still a useful property of the IssueInstance for informational reasons.

This makes the following changes to achieve the above:
* Creates new SharedTextKinds: `SINK_DETAIL` and `SOURCE_DETAIL`. These store concrete leaf names.
* The existing `SINK` and `SOURCE` SharedTextKinds will store leaf kinds instead.
* In this change, for back-compat reasons, the sink name/detail will be stored in both `SINK` and `SINK_DETAIL` (same for sources). Subsequent change will store it only in `SINK_DETAIL` after usages have been updated to read from the correct location.
* `IssueInstanceSharedTextAssoc` will contain the source/sink callable (name) associated with the issue instance along with the source/sink kinds.
* `TraceFrameSharedTextAssoc` will contain only the source/sink kinds associated with the trace frame.

Reviewed By: fahndrich

Differential Revision: D15444031

fbshipit-source-id: 6195eeebe7b926198a8e161da37c6ae93074eb04