Add TypedDict 'copy constructor'
Summary:
MyPy has an undocumented "copy constructor" functionality where you can construct a TypedDict by passing a compliant dictionary literal in as a single argument.  This implements that as an overload on the constructor.
This is arguably an anti-pattern, as you should be able to achieve the same result with
```
x: TD = { ... }
```
without the runtime cost of calling this identity function.

For now, we want to be as permissive as possible to ease the rollout of TypedDicts.

Reviewed By: sinancepel

Differential Revision: D15523232

fbshipit-source-id: 492302573c6b1716b5a3d35fb710f8b2e839980f