Correct handling of Optionals in Type.union
Summary:
For some reason D6094721, when trying to  normalize `Union[T, Optional[T], ...]` to `Union[Optional[T], ...]`, limited the transformation to only `Primitive` Ts.
I'm not sure why this was done, but this was causing strange downstream behavior.
The inciting problem was that `join`ing anything against a type variable returns `Type.union left right`.  With the incomplete implementation, this meant joining T and Optional[T] gave you Optional[Union[T, T]].  After going through an assert, this was left as Union[T, T].  This then ran into another simplifying assumption, that claimed that nothing was less or equal to a type variable except itself and Bottom (and sometimes Any).  This assumption is violated when we have Union[T, T].  It seems reasonable to assume these kinds of unions don't exist.

Reviewed By: sinancepel

Differential Revision: D15531543

fbshipit-source-id: 30641aee6291360c63e13477cc69ed47a112baae