streamline namedTuples tests
Summary:
They weren't being preprocessed, which is not good for writing new tests. Adds the preprocessing.

This caught a few issues in the implementation of namedTuples - this diff reintroduces the check that T42893621 was handling - dynamic class registration is not something Pyre is architected to handle, and fixes how we add the name for the class parameter.

Reviewed By: shannonzhu

Differential Revision: D15523538

fbshipit-source-id: d4c8a98880036be24f7bdf17d6f2af379367c011