Define __getitem__ in GenericMeta instead of Generic
Summary:
Failing case before:
```
x: typing.Iterable[int]
x[0]
```

Would return:
```
Too many arguments [19]: Call `typing.Generic.__getitem__` expects 0 positional arguments, 1 was provided.
```

This should error that getitem is not a method of Iterable!
It was erroring this way because we were accessing the class method manually added to typing.Generic as an instance method, stripping away the first parameter in the process. What we really want is to only implement getitem in GenericMeta, which is useful when some class inheriting from Generic is directly called like `Generic[]` or `Foo[]` because we need to support this syntax for type variables. An instance of Generic or one of its subclasses should *not* know about any getitem.

Reviewed By: mrkmndz

Differential Revision: D15523205

fbshipit-source-id: 8d9168f85aa8dcca16083657834c41c55d9aa8dd