Make type constraint related tests more readable
Summary:
As promised as a follow up from D15339143, this diff presents a more comprehensible interface for the tests for type constraints.
This is done through using the polymorphic variants to carry instances of the `Type.Variable.pair` variant which means we can pass both kinds of variables in through the same function parameter.
This also takes advantage of our implemented parameter variadic declaration functionality to allow the typeorder tests to parse string versions of the relevant types.

Reviewed By: shannonzhu

Differential Revision: D15527919

fbshipit-source-id: 49158b465e7475c8d54524b379566ecc415b54dd