remove interactive build prompt when running Pyre for unbuilt targets
Summary:
Right now, when a user attempts to run Pyre in project with buck, they get prompted for whether to build buck targets if they're missing. There is almost never a case where we want to not run Pyre when targets are unbuilt.
Refactored tools/pyre/client/buck.py so that there is no such prompt. Since the `prompt` parameter is now unused, they are all removed from codebase.

Reviewed By: sinancepel

Differential Revision: D15562284

fbshipit-source-id: 3204f32528de152bc0164e787f358754bcb7798d